# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Evakuasi Tsunami Padang (Offline-First Backend)",
    "description": "Backend untuk sinkronisasi data prabencana dan simulasi bottleneck.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/shelter/checkin": {
      "post": {
        "tags": [
          "Shelters"
        ],
        "summary": "Shelter Checkin",
        "operationId": "shelter_checkin_api_v1_shelter_checkin_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckInResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shelter/occupancy": {
      "post": {
        "tags": [
          "Shelters"
        ],
        "summary": "Report Occupancy",
        "operationId": "report_occupancy_api_v1_shelter_occupancy_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OccupancyReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OccupancyStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shelter/{external_id}/occupancy": {
      "get": {
        "tags": [
          "Shelters"
        ],
        "summary": "Get Occupancy Status",
        "operationId": "get_occupancy_status_api_v1_shelter__external_id__occupancy_get",
        "parameters": [
          {
            "name": "external_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "External Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OccupancyStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckInRequest": {
        "properties": {
          "event_external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event External Id",
            "description": "External ID kejadian; kosong berarti event aktif terbaru"
          },
          "evacuation_point_external_id": {
            "type": "string",
            "title": "Evacuation Point External Id",
            "description": "External ID Tempat Evakuasi (TES/TEA)"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude",
            "description": "Garis lintang lokasi pengguna"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude",
            "description": "Garis bujur lokasi pengguna"
          },
          "accuracy_m": {
            "type": "number",
            "title": "Accuracy M",
            "description": "Akurasi GPS (maksimal 35m)"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "description": "Status keselamatan pengguna",
            "default": "Selamat"
          }
        },
        "type": "object",
        "required": [
          "evacuation_point_external_id",
          "latitude",
          "longitude",
          "accuracy_m"
        ],
        "title": "CheckInRequest"
      },
      "CheckInResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "event_external_id": {
            "type": "string",
            "title": "Event External Id"
          },
          "evacuation_point_external_id": {
            "type": "string",
            "title": "Evacuation Point External Id"
          },
          "checked_in_at": {
            "type": "string",
            "format": "date-time",
            "title": "Checked In At"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message",
          "event_external_id",
          "evacuation_point_external_id",
          "checked_in_at"
        ],
        "title": "CheckInResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OccupancyReportRequest": {
        "properties": {
          "evacuation_point_external_id": {
            "type": "string",
            "title": "Evacuation Point External Id"
          },
          "level": {
            "type": "string",
            "pattern": "^(LOW|MODERATE|FULL)$",
            "title": "Level"
          }
        },
        "type": "object",
        "required": [
          "evacuation_point_external_id",
          "level"
        ],
        "title": "OccupancyReportRequest"
      },
      "OccupancyStatusResponse": {
        "properties": {
          "evacuation_point_external_id": {
            "type": "string",
            "title": "Evacuation Point External Id"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "report_count": {
            "type": "integer",
            "title": "Report Count"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "Laporan pengguna yang sudah check-in"
          }
        },
        "type": "object",
        "required": [
          "evacuation_point_external_id",
          "level",
          "report_count"
        ],
        "title": "OccupancyStatusResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}